requirements.txt
Flask==3.1.0
Werkzeug==3.1.0
flask-login==0.6.3
flask-wtf==1.2.2
python-dotenv==1.0.1
qrcode[pil]==8.0
pyotp==2.9.0
Pillow==11.1.0
markupsafe==3.0.2
gunicorn==23.0.0
